torii.lib: Moved `torii.lib.soc.wishbone` into `torii.lib.bus.wishbone`
This moves the wishbone bus constructs into  the new `torii.lib.bus` module where we plan to centralize all our busses, and will be where the AMBA5 busses land as well.

This lets gateware that wants to use an internal bus, but is not quite an SoC not bother with the SoC library

This also adds the deprecation stubs for `torii.lib.soc.wishbone`.
