+++ docs/library/bus/index.md
@@ -1,0 +1,12 @@
+# Busses
+
+```{eval-rst}
+.. automodule:: torii.lib.bus
+
+```
+
+```{toctree}
+:maxdepth: 4
+
+wishbone
+```
