+++ docs/library/soc/index.md
@@ -9,6 +9,5 @@ event
 periph
 
 csr/index
-wishbone/index
 
 ```
