--- docs/library/soc/wishbone/index.md
@@ -1,11 +1,0 @@
-# Wishbone Bus
-
-```{eval-rst}
-
-.. autoclass:: torii.lib.soc.wishbone.bus.CycleType
-.. autoclass:: torii.lib.soc.wishbone.bus.BurstTypeExt
-.. autoclass:: torii.lib.soc.wishbone.bus.Interface
-.. autoclass:: torii.lib.soc.wishbone.bus.Decoder
-.. autoclass:: torii.lib.soc.wishbone.bus.Arbiter
-
-```
