+++ torii/lib/bus/__init__.py
@@ -1,0 +1,14 @@
+# SPDX-License-Identifier: BSD-2-Clause
+
+'''
+
+.. _Wishbone: https://opencores.org/howto/wishbone
+
+The :py:mod:`torii.lib.bus` module provides industry standard bus interfaces, crossbars, and other
+utilities for busses such as `Wishbone`_.
+
+'''
+
+__all__ = (
+
+)
